- Changed CoinSelectComponent handling to Dictionary GridStack intersections.

+++ Assets/Scripts/Grid.cs
@@ -2,12 +2,13 @@
 using System;
 using System.Collections;
 
+
 public class Point
 {
-	public float x;
-	public float y;
+	public int x;
+	public int y;
 
-	public Point(float x, float y)
+	public Point(int x, int y)
 	{
 		this.x = x;
 		this.y = y;
@@ -18,6 +19,7 @@ public class Point
 		return "x:" + this.x + ", y:" + this.y;
 	}
 }
+
 
 public class Grid
 {
